Display warning message for unsupported browsers
This adds Modernizr to check browser features and conditionally display
a warning message directing users to use a modern version of Firefox or
Chrome. The specific features checked are:

- opacity
- canvas
- SVG
- JSON
- WebGL

The Modernizr lib was generated with only these features selected, along
with class insertion with 'mod-' as the class prefix. The es5-strictmode
feature, though import to CATMAID, was not used because it wrongly flags
some strictmode-compliant browsers in Modernizr 2.8.3. Other ES5 feature
checks are scheduled to be added in Modernizr 3.

The warning message modally locks out the interface for unsupported
browsers. Browsers supporting all required features except WebGL (IE9,
IE10) are given the option of dismissing the message, since most
browsing and tracing functionality is present. IE11 is fully compliant
and does not display a warning.

Tested on IE6 and IE7 on WinXP, IE8 through IE11 on Win7, and current
Chrome and Firefox on Ubuntu 14.04 and OSX 10.9.

Closes acardona/CATMAID#557.