Stack: order layers
Add a drawing order to layers in stack. Specify this order in the
layerOrders property of the stack as an array of keys of layers ordered
back to front. Layers are reordered using the moveLayer method of stack.
Layer drawing based on DOM order is acheived using a container div
inside the stack view, accessible via Stack.getLayersView. Layers which
do not place their view inside this layer are not reorderable.

For layers with non-DOM ordering, such as PixiTileLayer, an optional
notifyReorder method can be specified on a layer, to let it handle
internal reordering.

Currently no UI is provided to reorder layers, although layer order is
reflected in TileLayerControl.